AI News Summarizer/requirements.txt
openai==0.27.8
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
